Fix edge case for posts without subheadings
Previously, if a post didn't have a subheading, there would be a match
error in getting the subheading and the app would crash. Instead of
using pattern matching when running Regex, I instead used a case
statement later in the function.

If the subheading was nil, as would be the case if no subheading was
used, it gets set to an empty string. Otherwise, we use pattern matching
to get just the returned subheading and not the regex.