Refactor use of defaults for blog_posts
Previously, the `blog_posts` functions were using defaults by declaring
them in the function arguments, i.e., `blog_posts(characters \\ 304)`.
This was originally fine as there was only one default being called.
With the addition of a default for returning the raw HTML, this needed
to be changed to avoid issues when only one of the two defaults was
being changed in the function call.